refactor : util 패키지 분리 및 StringConcatenater 클래스 변경
