+++ src/test/java/lotto/domain/StringConcatenaterTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.util.StringConcatenater;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
